modify BSR code according comments

+++ openair2/LAYER2/MAC/config.c
@@ -233,7 +233,7 @@ rrc_mac_config_req(
       if (mac_MainConfig->ext1 && mac_MainConfig->ext1->sr_ProhibitTimer_r9) {
         UE_mac_inst[Mod_idP].scheduling_info.sr_ProhibitTimer  = (uint16_t) *mac_MainConfig->ext1->sr_ProhibitTimer_r9;
       } else {
-        UE_mac_inst[Mod_idP].scheduling_info.sr_ProhibitTimer  = (uint16_t) 0;
+        UE_mac_inst[Mod_idP].scheduling_info.sr_ProhibitTimer  = 0;
       }
 
       if (mac_MainConfig->ext2 && mac_MainConfig->ext2->mac_MainConfig_v1020) {
